Biblio_Api/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
marshmallow==3.20.1
python-dotenv==1.0.0
Werkzeug==2.3.7
PyJWT==2.8.0